.ipynb_checkpoints/requirements-checkpoint.txt
torchvision==0.8.0
matplotlib
Pillow
Flask==1.0.3
waitress